+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+## [2.9.33] - 2023-11-08
 ### Added
 - support for enhanced gender options from Shopgate
 
@@ -216,7 +217,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## Version 2.9.0
 - Created Initial Plugin
 
-[Unreleased]: https://github.com/shopgate/cart-integration-magento2-base/compare/2.9.32...HEAD
+[Unreleased]: https://github.com/shopgate/cart-integration-magento2-base/compare/2.9.33...HEAD
+[2.9.33]: https://github.com/shopgate/cart-integration-magento2-base/compare/2.9.32...2.9.33
 [2.9.32]: https://github.com/shopgate/cart-integration-magento2-base/compare/2.9.31...2.9.32
 [2.9.31]: https://github.com/shopgate/cart-integration-magento2-base/compare/2.9.30...2.9.31
 [2.9.30]: https://github.com/shopgate/cart-integration-magento2-base/compare/2.9.29...2.9.30